design mainPage
