+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:dsc_solution_challenge_2020/loginPage.dart';
 import 'package:dsc_solution_challenge_2020/registerPage.dart';
 import 'package:flutter/material.dart';
 import 'package:dsc_solution_challenge_2020/mainPage.dart';
+import 'package:dsc_solution_challenge_2020/signInPage.dart';
 
 void main() => runApp(MyApp());
 
@@ -15,11 +16,12 @@ class MyApp extends StatelessWidget {
       
         primarySwatch: Colors.blue,
       ),
-      initialRoute: LoginPage.id,
+      initialRoute: MainPage.id,
+      //initialRoute: LoginPage.id,
       routes: {
         MainPage.id: (context) => MainPage(),
         LoginPage.id: (context) => LoginPage(),
-        RegisterPage.id: (context) => RegisterPage(),
+        SignInPage.id: (context) => SignInPage(),
         
       },
     );
