+++ lib/models/profile.dart
@@ -5,7 +5,7 @@ class Profile{
   final String name;
   final String address;
   final int age;
-  final Image photo;
+  final ImageProvider photo;
   final String comments;
   final String phoneNumber;
   //보호자 이름, 전화번호 등등등등등
